requirements.txt
streamlit==1.31.0
pandas==2.2.0
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
azure-openai==0.6.0
serpapi==0.1.2
requests==2.31.0
google-search-results==2.24